pyproject.toml
[tool.poetry]
name = "dataset downloader"
version = "0.1.0"
description = ""
authors = ["shaheer-2000 <shaheer.2000@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydrive = "^1.3.1"
youtube-dl = "^2021.12.17"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
